# docs/openapi/asstrack.v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "AssTrack.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/system/seed": {
      "post": {
        "tags": [
          "AssTrack.Api"
        ],
        "summary": "Seed demo data. Use reset=true to wipe and re-seed seeded records only.",
        "operationId": "SeedDemoData",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/system/maintenance/clean-null-island": {
      "post": {
        "tags": [
          "AssTrack.Api"
        ],
        "summary": "Remove historical observations near 0,0 and reset geofence state for affected devices.",
        "operationId": "CleanNullIslandObservationNoise",
        "parameters": [
          {
            "name": "dryRun",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/system/maintenance/clean-auto-created-provider-assets": {
      "post": {
        "tags": [
          "AssTrack.Api"
        ],
        "summary": "Detach devices from previously auto-created provider assets and remove those asset records.",
        "operationId": "CleanAutoCreatedProviderAssets",
        "parameters": [
          {
            "name": "dryRun",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SeedRequest": {
        "type": "object",
        "properties": {
          "reset": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "API key authentication",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    { }
  ],
...
}